[Silver III] Title: 주유소, Time: 832 ms, Memory: 63936 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.kt
@@ -19,7 +19,7 @@ fun main() {
 
             remaining += assigned
 
-            price * assigned
+            price.toLong() * assigned
         }
 
     println(result)
